refactor proposal delivery routes
