+++ src/app/router_registry.py
@@ -28,6 +28,7 @@ from app.routers.foundation import router as foundation_router
 from app.routers.intake import router as intake_router
 from app.routers.platform import router as platform_router
 from app.routers.portfolio import router as portfolio_router
+from app.routers.proposal_delivery import router as proposal_delivery_router
 from app.routers.proposal_memos import router as proposal_memos_router
 from app.routers.proposal_narratives import router as proposal_narratives_router
 from app.routers.proposal_operations import router as proposal_operations_router
@@ -56,6 +57,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(proposal_versions_router)
     app.include_router(proposal_workflow_router)
     app.include_router(proposal_narratives_router)
+    app.include_router(proposal_delivery_router)
     app.include_router(proposal_memos_router)
     app.include_router(platform_router)
     app.include_router(domain_products_router)
